Memperbaiki tampilan dan merapikan direktori
