+++ app/Http/Controllers/JasaPrint/JasaPrintController.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\JasaPrint;
 
-use App\Software;
+use App\Http\Controllers\Controller;
+use App\JasaPrint;
 use Illuminate\Http\Request;
 
-class SoftwareController extends Controller
+class JasaPrintController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -41,10 +42,10 @@ class SoftwareController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Software  $software
+     * @param  \App\JasaPrint  $jasaPrint
      * @return \Illuminate\Http\Response
      */
-    public function show(Software $software)
+    public function show(JasaPrint $jasaPrint)
     {
         //
     }
@@ -52,10 +53,10 @@ class SoftwareController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Software  $software
+     * @param  \App\JasaPrint  $jasaPrint
      * @return \Illuminate\Http\Response
      */
-    public function edit(Software $software)
+    public function edit(JasaPrint $jasaPrint)
     {
         //
     }
@@ -64,10 +65,10 @@ class SoftwareController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Software  $software
+     * @param  \App\JasaPrint  $jasaPrint
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Software $software)
+    public function update(Request $request, JasaPrint $jasaPrint)
     {
         //
     }
@@ -75,10 +76,10 @@ class SoftwareController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Software  $software
+     * @param  \App\JasaPrint  $jasaPrint
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Software $software)
+    public function destroy(JasaPrint $jasaPrint)
     {
         //
     }
