+++ app/Http/Controllers/MataKuliah/MataKuliahController.php
@@ -1,0 +1,86 @@
+<?php
+
+namespace App\Http\Controllers\MataKuliah;
+
+use App\Http\Controllers\Controller;
+use App\MataKuliah;
+use Illuminate\Http\Request;
+
+class MataKuliahController extends Controller
+{
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function index()
+    {
+        //
+    }
+
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function create()
+    {
+        //
+    }
+
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function store(Request $request)
+    {
+        //
+    }
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  \App\MataKuliah  $mataKuliah
+     * @return \Illuminate\Http\Response
+     */
+    public function show(MataKuliah $mataKuliah)
+    {
+        //
+    }
+
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  \App\MataKuliah  $mataKuliah
+     * @return \Illuminate\Http\Response
+     */
+    public function edit(MataKuliah $mataKuliah)
+    {
+        //
+    }
+
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  \App\MataKuliah  $mataKuliah
+     * @return \Illuminate\Http\Response
+     */
+    public function update(Request $request, MataKuliah $mataKuliah)
+    {
+        //
+    }
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  \App\MataKuliah  $mataKuliah
+     * @return \Illuminate\Http\Response
+     */
+    public function destroy(MataKuliah $mataKuliah)
+    {
+        //
+    }
+}
