+++ app/Http/Controllers/TransaksiJasaPrint/TransaksiJasaPrintController.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\TransaksiJasaPrint;
 
-use App\TransaksiJasaInstallasi;
+use App\Http\Controllers\Controller;
+use App\TransaksiJasaPrint;
 use Illuminate\Http\Request;
 
-class TransaksiJasaInstallasiController extends Controller
+class TransaksiJasaPrintController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -41,10 +42,10 @@ class TransaksiJasaInstallasiController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\TransaksiJasaInstallasi  $transaksiJasaInstallasi
+     * @param  \App\TransaksiJasaPrint  $transaksiJasaPrint
      * @return \Illuminate\Http\Response
      */
-    public function show(TransaksiJasaInstallasi $transaksiJasaInstallasi)
+    public function show(TransaksiJasaPrint $transaksiJasaPrint)
     {
         //
     }
@@ -52,10 +53,10 @@ class TransaksiJasaInstallasiController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\TransaksiJasaInstallasi  $transaksiJasaInstallasi
+     * @param  \App\TransaksiJasaPrint  $transaksiJasaPrint
      * @return \Illuminate\Http\Response
      */
-    public function edit(TransaksiJasaInstallasi $transaksiJasaInstallasi)
+    public function edit(TransaksiJasaPrint $transaksiJasaPrint)
     {
         //
     }
@@ -64,10 +65,10 @@ class TransaksiJasaInstallasiController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\TransaksiJasaInstallasi  $transaksiJasaInstallasi
+     * @param  \App\TransaksiJasaPrint  $transaksiJasaPrint
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, TransaksiJasaInstallasi $transaksiJasaInstallasi)
+    public function update(Request $request, TransaksiJasaPrint $transaksiJasaPrint)
     {
         //
     }
@@ -75,10 +76,10 @@ class TransaksiJasaInstallasiController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\TransaksiJasaInstallasi  $transaksiJasaInstallasi
+     * @param  \App\TransaksiJasaPrint  $transaksiJasaPrint
      * @return \Illuminate\Http\Response
      */
-    public function destroy(TransaksiJasaInstallasi $transaksiJasaInstallasi)
+    public function destroy(TransaksiJasaPrint $transaksiJasaPrint)
     {
         //
     }
