+++ routes/web.php
@@ -22,3 +22,5 @@ Route::resource('Account', 'Account\AccountController')->except('create');
 Route::resource('PeminjamanLab', 'PeminjamanLab\PeminjamanLabController')->except('show');
 Route::resource('PeminjamanAlat', 'PeminjamanAlat\PeminjamanAlatController')->except('show');
 Route::resource('SuratBebasLabkom', 'SuratBebasLabkom\SuratBebasLabkomController')->except('show');
+Route::resource('JasaInstallasi', 'JasaInstallasi\JasaInstallasiController')->except('show');
+Route::resource('JasaPrint', 'JasaPrint\JasaPrintController')->except('show');
